Stock data analysis module 1 VBA code
Module 1 contains the VBA code for the stock market data analysis that iterates through each year worksheet in the workbook and calculates the stock price change for the year, the percent change for the year, and the total volume for the year and puts those values in a table next to the raw data. This module uses a for each worksheet function to iterate through all worksheets, a for loop to iterate through all the stock symbols on the worksheet, and multiple conditional statements to determine the stock price change, percent change, and total volume for the specified stock. This module also contains conditional formatting to mark the cells that have a positive percent change for the year green and ones that have a negative percent change red.