Stock data analysis module 2
This module uses the for each worksheet function to iterate through all worksheets in the workbook. It also uses a nested for loop where the outside loop iterates through the percent change output column and the total volume output column. Conditionals are used within the for loops to capture the greatest percent increase, greatest percent decrease, and greatest total volume for each year of stock data. These values are entered into a table to the right of the data.